player-item-persist: add faithful SmTitleInfo(bool)/(Player,bool) overloads+writeImpl cases 4/5 (NpcFactions), HouseObject SetXYZH (byte)heading, ChatCommand out sbyte ACCESS_LEVELS, GeneralInstancePosition TeleportTo (byte)h (119->113)

+++ dotnetConversion/src/Aion.GameServer/Model/GameObjects/HouseObject.cs
@@ -177,7 +177,7 @@ public abstract class HouseObject : VisibleObject, IExpirable, IPersistable
             this.heading = heading;
             SetPersistentState(IPersistable.PersistentState.UPDATE_REQUIRED);
             if (Position != null)
-                Position.SetXYZH(null, null, null, heading);
+                Position.SetXYZH(null, null, null, (byte)heading);
         }
     }
